refactor(model.event.EventType) : 이벤트 종류를 구분하는 Enum 클래스 이름 변경
