+++ src/main/java/christmas/model/event/ChampagneGiftEvent.java
@@ -1,14 +1,14 @@
 package christmas.model.event;
 
 import static christmas.model.Menu.CHAMPAGNE;
-import static christmas.model.event.EventType.CHAMPAGNE_GIFT;
+import static christmas.model.event.EventKind.CHAMPAGNE_GIFT;
 
 import christmas.model.ClientOrders;
 import christmas.model.GiftOrders;
 import christmas.model.Menu;
 
 public class ChampagneGiftEvent implements GiftEvent {
-    private static final EventType type = CHAMPAGNE_GIFT;
+    private static final EventKind type = CHAMPAGNE_GIFT;
     private static final int MIN_ORDER_AMOUNT_FOR_SATISFIED_CONDITION = 120_000;
     private static final Menu GIFT_MENU = CHAMPAGNE;
     private static final int GIFT_COUNT = 1;
@@ -26,7 +26,7 @@ public class ChampagneGiftEvent implements GiftEvent {
     }
 
     @Override
-    public EventType getType() {
+    public EventKind getType() {
         return type;
     }
 }
