+++ src/main/java/christmas/model/event/DecemberBadgeEvent.java
@@ -1,13 +1,13 @@
 package christmas.model.event;
 
-import static christmas.model.event.EventType.DECEMBER_BADGE;
+import static christmas.model.event.EventKind.DECEMBER_BADGE;
 
 import christmas.model.Badge;
 import christmas.model.DecemberBadge;
 import java.util.Optional;
 
 public class DecemberBadgeEvent implements BadgeEvent {
-    private final static EventType type = DECEMBER_BADGE;
+    private final static EventKind type = DECEMBER_BADGE;
 
     @Override
     public Optional<Badge> apply(final int benefitAmount) {
@@ -15,7 +15,7 @@ public class DecemberBadgeEvent implements BadgeEvent {
     }
 
     @Override
-    public EventType getType() {
+    public EventKind getType() {
         return type;
     }
 }
