+++ src/main/java/christmas/model/event/DiscountResult.java
@@ -4,9 +4,9 @@ import java.util.EnumMap;
 import java.util.Map;
 
 public class DiscountResult {
-    private final Map<EventType, Integer> result = new EnumMap<>(EventType.class);
+    private final Map<EventKind, Integer> result = new EnumMap<>(EventKind.class);
 
-    public void addResult(EventType eventType, int discountAmount) {
-        result.put(eventType, discountAmount);
+    public void addResult(EventKind eventKind, int discountAmount) {
+        result.put(eventKind, discountAmount);
     }
 }
